requirements.txt
streamlit==1.52.1
langchain-google-genai==4.0.0
python-dotenv==1.2.1
PyPDF2==3.0.1